Refactor runtime platform handling
