+++ runtime/runtime.go
@@ -18,6 +18,7 @@ package runtime
 
 import (
 	"ballerina-lang-go/bir"
+	"ballerina-lang-go/pal"
 	"ballerina-lang-go/runtime/internal/exec"
 	"ballerina-lang-go/runtime/internal/modules"
 	"ballerina-lang-go/semtypes"
@@ -28,7 +29,7 @@ import (
 // and is used as the execution context for interpreting BIR packages.
 type Runtime struct {
 	registry *modules.Registry
-	platform Platform
+	platform pal.Platform
 }
 
 // ModuleInitializer is a function that can install modules (e.g. stdlibs) into
@@ -39,7 +40,7 @@ var moduleInitializers []ModuleInitializer
 
 // NewRuntime constructs a new runtime with an empty registry and runs all
 // registered module initializers.
-func NewRuntime(platform Platform) *Runtime {
+func NewRuntime(platform pal.Platform) *Runtime {
 	rt := &Runtime{
 		registry: modules.NewRegistry(),
 		platform: platform,
@@ -51,7 +52,7 @@ func NewRuntime(platform Platform) *Runtime {
 }
 
 // Platform returns the platform configuration of this runtime instance.
-func (rt *Runtime) Platform() Platform {
+func (rt *Runtime) Platform() pal.Platform {
 	return rt.platform
 }
